+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "silly-bot",
-  "version": "1.0.13",
+  "version": "1.0.14",
   "description": "A little discord bot project",
   "main": "index.js",
   "scripts": {
