+++ src/types/guild_member.ts
@@ -1,0 +1,5 @@
+// User template
+export interface template_user {
+    id: string,
+    points: number,
+};
